backend/requirements.txt
fastapi
uvicorn
line-bot-sdk
firebase-admin
python-dotenv
requests
pytest
pytest-mock
